+++ src/EditorFeatures/Core/InlineRename/AbstractEditorInlineRenameService.cs
@@ -76,7 +76,7 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.InlineRename
             // If rename is invoked on a member group reference in a nameof expression, then the
             // RenameOverloads option should be forced on.
             var forceRenameOverloads = tokenRenameInfo.IsMemberGroup;
-            var symbol = await RenameLocations.ReferenceProcessing.TryGetRenamableSymbolAsync(document, triggerToken.SpanStart, cancellationToken: cancellationToken).ConfigureAwait(false);
+            var symbol = await RenameUtilities.TryGetRenamableSymbolAsync(document, triggerToken.SpanStart, cancellationToken: cancellationToken).ConfigureAwait(false);
             if (symbol == null)
                 return new FailureInlineRenameInfo(EditorFeaturesResources.You_cannot_rename_this_element);
