+++ src/Workspaces/Core/Portable/Rename/ConflictEngine/ConflictResolver.cs
@@ -79,36 +79,40 @@ namespace Microsoft.CodeAnalysis.Rename.ConflictEngine
                 }
             }
 
-            var renameLocations = await lightweightRenameLocations.ToRenameLocationsAsync(cancellationToken).ConfigureAwait(false);
-            if (renameLocations is null)
+            var heavyweightLocations = await lightweightRenameLocations.ToHeavyweightAsync(cancellationToken).ConfigureAwait(false);
+            if (heavyweightLocations is null)
                 return new ConflictResolution(WorkspacesResources.Failed_to_resolve_rename_conflicts);
 
-            return await ResolveConflictsInCurrentProcessAsync(
-                renameLocations, replacementText, nonConflictSymbols, cancellationToken).ConfigureAwait(false);
+            return await ResolveHeavyweightConflictsInCurrentProcessAsync(
+                heavyweightLocations, replacementText, nonConflictSymbols, cancellationToken).ConfigureAwait(false);
         }
 
-        internal static async Task<ConflictResolution> ResolveConflictsInCurrentProcessAsync(
-            RenameLocations renameLocationSet,
+        /// <summary>
+        /// Finds any conflicts that would arise from using <paramref name="replacementText"/> as the new name for a
+        /// symbol and returns how to resolve those conflicts.  Will not cross any process boundaries to do this.
+        /// </summary>
+        internal static async Task<ConflictResolution> ResolveHeavyweightConflictsInCurrentProcessAsync(
+            HeavyweightRenameLocations renameLocations,
             string replacementText,
             ImmutableHashSet<ISymbol>? nonConflictSymbols,
             CancellationToken cancellationToken)
         {
             // when someone e.g. renames a symbol from metadata through the API (IDE blocks this), we need to return
-            var renameSymbolDeclarationLocation = renameLocationSet.Symbol.Locations.Where(loc => loc.IsInSource).FirstOrDefault();
+            var renameSymbolDeclarationLocation = renameLocations.Symbol.Locations.Where(loc => loc.IsInSource).FirstOrDefault();
             if (renameSymbolDeclarationLocation == null)
             {
                 // Symbol "{0}" is not from source.
-                return new ConflictResolution(string.Format(WorkspacesResources.Symbol_0_is_not_from_source, renameLocationSet.Symbol.Name));
+                return new ConflictResolution(string.Format(WorkspacesResources.Symbol_0_is_not_from_source, renameLocations.Symbol.Name));
             }
 
             var resolution = await ResolveMutableConflictsAsync(
-                renameLocationSet, renameSymbolDeclarationLocation, replacementText, nonConflictSymbols, cancellationToken).ConfigureAwait(false);
+                renameLocations, renameSymbolDeclarationLocation, replacementText, nonConflictSymbols, cancellationToken).ConfigureAwait(false);
 
             return resolution.ToConflictResolution();
         }
 
         private static Task<MutableConflictResolution> ResolveMutableConflictsAsync(
-            RenameLocations renameLocationSet,
+            HeavyweightRenameLocations renameLocationSet,
             Location renameSymbolDeclarationLocation,
             string replacementText,
             ImmutableHashSet<ISymbol>? nonConflictSymbols,
