+++ src/Workspaces/Core/Portable/Rename/HeavyweightRenameLocations.ReferenceProcessing.cs
@@ -4,8 +4,10 @@
 
 using System;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Linq;
+using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.FindSymbols;
@@ -14,9 +16,6 @@ using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Shared.Extensions;
 using Microsoft.CodeAnalysis.Text;
 using Roslyn.Utilities;
-using System.Text.RegularExpressions;
-using System.Collections.Immutable;
-using System.Text;
 
 namespace Microsoft.CodeAnalysis.Rename
 {
@@ -24,101 +23,10 @@ namespace Microsoft.CodeAnalysis.Rename
     /// A helper class that contains some of the methods and filters that must be used when
     /// processing the raw results from the FindReferences API.
     /// </summary>
-    internal sealed partial class RenameLocations
+    internal sealed partial class HeavyweightRenameLocations
     {
         internal static class ReferenceProcessing
         {
-            /// <summary>
-            /// Given a symbol in a document, returns the "right" symbol that should be renamed in
-            /// the case the name binds to things like aliases _and_ the underlying type at once.
-            /// </summary>
-            public static async Task<ISymbol?> TryGetRenamableSymbolAsync(
-                Document document, int position, CancellationToken cancellationToken)
-            {
-                var symbol = await SymbolFinder.FindSymbolAtPositionAsync(document, position, cancellationToken: cancellationToken).ConfigureAwait(false);
-                if (symbol == null)
-                    return null;
-
-                var definitionSymbol = await FindDefinitionSymbolAsync(symbol, document.Project.Solution, cancellationToken).ConfigureAwait(false);
-                Contract.ThrowIfNull(definitionSymbol);
-
-                return definitionSymbol;
-            }
-
-            /// <summary>
-            /// Given a symbol, finds the symbol that actually defines the name that we're using.
-            /// </summary>
-            public static async Task<ISymbol> FindDefinitionSymbolAsync(
-                ISymbol symbol, Solution solution, CancellationToken cancellationToken)
-            {
-                Contract.ThrowIfNull(symbol);
-                Contract.ThrowIfNull(solution);
-
-                // Make sure we're on the original source definition if we can be
-                var foundSymbol = await SymbolFinder.FindSourceDefinitionAsync(
-                    symbol, solution, cancellationToken).ConfigureAwait(false);
-
-                var bestSymbol = foundSymbol ?? symbol;
-                symbol = bestSymbol;
-
-                // If we're renaming a property, it might be a synthesized property for a method
-                // backing field.
-                if (symbol.Kind == SymbolKind.Parameter)
-                {
-                    if (symbol.ContainingSymbol.Kind == SymbolKind.Method)
-                    {
-                        var containingMethod = (IMethodSymbol)symbol.ContainingSymbol;
-                        if (containingMethod.AssociatedSymbol is IPropertySymbol)
-                        {
-                            var associatedPropertyOrEvent = (IPropertySymbol)containingMethod.AssociatedSymbol;
-                            var ordinal = containingMethod.Parameters.IndexOf((IParameterSymbol)symbol);
-                            if (ordinal < associatedPropertyOrEvent.Parameters.Length)
-                            {
-                                return associatedPropertyOrEvent.Parameters[ordinal];
-                            }
-                        }
-                    }
-                }
-
-                // if we are renaming a compiler generated delegate for an event, cascade to the event
-                if (symbol.Kind == SymbolKind.NamedType)
-                {
-                    var typeSymbol = (INamedTypeSymbol)symbol;
-                    if (typeSymbol.IsImplicitlyDeclared && typeSymbol.IsDelegateType() && typeSymbol.AssociatedSymbol != null)
-                    {
-                        return typeSymbol.AssociatedSymbol;
-                    }
-                }
-
-                // If we are renaming a constructor or destructor, we wish to rename the whole type
-                if (symbol.Kind == SymbolKind.Method)
-                {
-                    var methodSymbol = (IMethodSymbol)symbol;
-                    if (methodSymbol.MethodKind is MethodKind.Constructor or
-                        MethodKind.StaticConstructor or
-                        MethodKind.Destructor)
-                    {
-                        return methodSymbol.ContainingType;
-                    }
-                }
-
-                // If we are renaming a backing field for a property, cascade to the property
-                if (symbol.Kind == SymbolKind.Field)
-                {
-                    var fieldSymbol = (IFieldSymbol)symbol;
-                    if (fieldSymbol.IsImplicitlyDeclared &&
-                        fieldSymbol.AssociatedSymbol.IsKind(SymbolKind.Property))
-                    {
-                        return fieldSymbol.AssociatedSymbol;
-                    }
-                }
-
-                // in case this is e.g. an overridden property accessor, we'll treat the property itself as the definition symbol
-                var property = await TryGetPropertyFromAccessorOrAnOverrideAsync(bestSymbol, solution, cancellationToken).ConfigureAwait(false);
-
-                return property ?? bestSymbol;
-            }
-
             private static async Task<bool> ShouldIncludeSymbolAsync(
                 ISymbol referencedSymbol, ISymbol originalSymbol, Solution solution, bool considerSymbolReferences, CancellationToken cancellationToken)
             {
@@ -213,48 +121,10 @@ namespace Microsoft.CodeAnalysis.Rename
                 }
             }
 
-            internal static async Task<ISymbol?> TryGetPropertyFromAccessorOrAnOverrideAsync(
-                ISymbol symbol, Solution solution, CancellationToken cancellationToken)
-            {
-                if (symbol.IsPropertyAccessor())
-                {
-                    return ((IMethodSymbol)symbol).AssociatedSymbol;
-                }
-
-                if (symbol.IsOverride && symbol.GetOverriddenMember() != null)
-                {
-                    var originalSourceSymbol = await SymbolFinder.FindSourceDefinitionAsync(
-                        symbol.GetOverriddenMember(), solution, cancellationToken).ConfigureAwait(false);
-
-                    if (originalSourceSymbol != null)
-                    {
-                        return await TryGetPropertyFromAccessorOrAnOverrideAsync(originalSourceSymbol, solution, cancellationToken).ConfigureAwait(false);
-                    }
-                }
-
-                if (symbol.Kind == SymbolKind.Method &&
-                    symbol.ContainingType.TypeKind == TypeKind.Interface)
-                {
-                    var methodImplementors = await SymbolFinder.FindImplementationsAsync(
-                        symbol, solution, cancellationToken: cancellationToken).ConfigureAwait(false);
-
-                    foreach (var methodImplementor in methodImplementors)
-                    {
-                        var propertyAccessorOrAnOverride = await TryGetPropertyFromAccessorOrAnOverrideAsync(methodImplementor, solution, cancellationToken).ConfigureAwait(false);
-                        if (propertyAccessorOrAnOverride != null)
-                        {
-                            return propertyAccessorOrAnOverride;
-                        }
-                    }
-                }
-
-                return null;
-            }
-
             private static async Task<bool> IsPropertyAccessorOrAnOverrideAsync(
                 ISymbol symbol, Solution solution, CancellationToken cancellationToken)
             {
-                var result = await TryGetPropertyFromAccessorOrAnOverrideAsync(
+                var result = await RenameUtilities.TryGetPropertyFromAccessorOrAnOverrideAsync(
                     symbol, solution, cancellationToken).ConfigureAwait(false);
                 return result != null;
             }
@@ -526,7 +396,7 @@ namespace Microsoft.CodeAnalysis.Rename
                 IEnumerable<Tuple<string, int, TextSpan>> renameStringsAndPositions,
                 ArrayBuilder<RenameLocation> renameLocations)
             {
-                var regex = GetRegexForMatch(renameText);
+                var regex = RenameUtilities.GetRegexForMatch(renameText);
                 foreach (var renameStringAndPosition in renameStringsAndPositions)
                 {
                     var renameString = renameStringAndPosition.Item1;
@@ -547,59 +417,6 @@ namespace Microsoft.CodeAnalysis.Rename
                         var renameLocation = new RenameLocation(matchLocation, document.Id, containingLocationForStringOrComment: containingSpan);
                         renameLocations.Add(renameLocation);
                     }
-                }
-            }
-
-            private static Regex GetRegexForMatch(string matchText)
-            {
-                var matchString = string.Format(@"\b{0}\b", matchText);
-                return new Regex(matchString, RegexOptions.CultureInvariant);
-            }
-
-            internal static string ReplaceMatchingSubStrings(
-                string replaceInsideString,
-                string matchText,
-                string replacementText,
-                ImmutableSortedSet<TextSpan>? subSpansToReplace = null)
-            {
-                if (subSpansToReplace == null)
-                {
-                    // We do not have already computed sub-spans to replace inside the string.
-                    // Get regex for matches within the string and replace all matches with replacementText.
-                    var regex = GetRegexForMatch(matchText);
-                    return regex.Replace(replaceInsideString, replacementText);
-                }
-                else
-                {
-                    // We are provided specific matches to replace inside the string.
-                    // Process the input string from start to end, replacing matchText with replacementText
-                    // at the provided sub-spans within the string for these matches.
-                    var stringBuilder = new StringBuilder();
-                    var startOffset = 0;
-                    foreach (var subSpan in subSpansToReplace)
-                    {
-                        Debug.Assert(subSpan.Start <= replaceInsideString.Length);
-                        Debug.Assert(subSpan.End <= replaceInsideString.Length);
-
-                        // Verify that provided sub-span has a match with matchText.
-                        if (replaceInsideString.Substring(subSpan.Start, subSpan.Length) != matchText)
-                            continue;
-
-                        // Append the sub-string from last match till the next match
-                        var offset = subSpan.Start - startOffset;
-                        stringBuilder.Append(replaceInsideString.Substring(startOffset, offset));
-
-                        // Append the replacementText
-                        stringBuilder.Append(replacementText);
-
-                        // Update startOffset to process the next match.
-                        startOffset += offset + subSpan.Length;
-                    }
-
-                    // Append the remaining of the sub-string within replaceInsideString after the last match. 
-                    stringBuilder.Append(replaceInsideString.Substring(startOffset));
-
-                    return stringBuilder.ToString();
                 }
             }
         }
